+++ .storybook/config.js
@@ -1,0 +1,11 @@
+import { configure } from '@kadira/storybook';
+
+function loadStories() {
+  require('../stories/conversation');
+  require('../stories/messages');
+  require('../stories/message');
+  require('../stories/typing')
+  // require as many stories as you need.
+}
+
+configure(loadStories, module);