+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.0.4] - 2016-04-13
+- No Issue: Added Storybook to test components (Paul.O'Flaherty)
+- No Issue: Removed Karma and phantom JS as the testing suite (conor.fennell)
+- No Issue: Added enzyme as the new testing suite (conor.fennell)
+- No Issue: Added the display of messages (Paul.O'Flaherty)
+- No Issue: Added the Typing component (conor.fennell)
+
 ## [0.0.3] - 2016-04-12
 - No Issue: Updated the webpack example to use the 0.0.2 version of chat-template (conor.fennell)